srsdu: allow 2.5 ms scheduling request period

+++ apps/units/flexible_du/du_high/du_high_config.h
@@ -218,8 +218,8 @@ struct du_high_unit_pucch_config {
 
   /// \brief \c SR period in milliseconds.
   /// Among all values given in \c periodicityAndOffset, part of \c \SchedulingRequestResourceConfig, TS 38.331,
-  /// these are the only ones supported. Values: {1, 2, 4, 8, 10, 16, 20, 40, 80, 160, 320}.
-  unsigned sr_period_msec = 20;
+  /// these are the only ones supported. Values: {1, 2, 2.5, 4, 5, 8, 10, 16, 20, 40, 80, 160, 320}.
+  float sr_period_msec = 20.0F;
 
   /// PUCCH F1 resource parameters.
   /// Number of symbols for PUCCH Format 1. Values {4, 14}.
